src/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Ihor Horobets <ihor@horobets.me>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^12.3.0"
requests = "^2.22.0"
pymongo = "^3.10.1"
matplotlib = "^3.1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
